+++ tests/testthat/test-graph_test_closure.R
@@ -128,10 +128,12 @@ test_that("Non-Bonferroni adjusted p-values are less than Bonferroni", {
 
   expect_true(
     all(
-      graph_test_closure(bonferroni_holm(rep(1/4, 4)),
-                         rep(.01, 4))$outputs$adjusted_p >=
+      graph_test_closure(
+        bonferroni_holm(4),
+        rep(.01, 4)
+      )$outputs$adjusted_p >=
         graph_test_closure(
-          bonferroni_holm(rep(1/4, 4)),
+          bonferroni_holm(4),
           rep(.01, 4),
           test_types = "h"
         )$outputs$adjusted_p
@@ -394,7 +396,7 @@ test_that("closure internal consistency", {
 test_that("parametric floating point errors", {
   # Parametric adjusted p-values and c-values are now rounded to 10 decimals to
   # avoid floating point errors. These tests fail without this rounding.
-  bh <- bonferroni_holm(rep(1 / 3, 3))
+  bh <- bonferroni_holm(3)
   p <- rep(.025, 3)
 
   t_corr <- list(matrix(1, 3, 3))